refactor(application): split generate_sbom mod.rs into responsibility submodules
Split the 875-line GenerateSbomUseCase impl block into responsibility-scoped
submodules under src/application/use_cases/generate_sbom/:

- mod.rs: struct definition, new(), and execute() orchestration only (177 lines)
- filtering.rs: read_and_report_lockfile, apply_exclusion_filters,
  apply_group_filter, analyze_dependencies_if_requested
- checks.rs: check_vulnerabilities_if_requested, check_license_compliance_if_requested,
  check_abandoned_if_requested, check_non_pypi_if_requested,
  check_python_compatibility_if_requested, build_threshold_config,
  fetch_license_info, resolve_direct_names
- upgrade.rs: advise_upgrades_if_requested
- response.rs: build_response, build_dry_run_response

All method bodies moved verbatim (no logic changes); methods called from
execute() or tests.rs are pub(super), methods used only within their own
submodule stay private fn.

Deviates from the issue's literal grouping in three places, documented here:
- analyze_dependencies_if_requested moved to filtering.rs (not upgrade.rs):
  its DependencyGraph output feeds four other checks plus the response
  builder, not just the upgrade advisor, so it belongs with the other
  input-preparation steps.
- fetch_license_info (unassigned by the issue) moved to checks.rs: it shares
  LicenseRepository/EnrichedPackage with check_license_compliance_if_requested.
- resolve_direct_names moved to checks.rs (not filtering.rs): all three call
  sites are check methods that landed there.

tests.rs gained two import lines (EnrichedPackage, UpgradeRecommendation,
ThresholdConfig, VulnerabilityCheckResult) since `use super::*` no longer
re-exports these once they leave mod.rs's import list; no test logic changed.

All 865 lib tests pass unchanged. cargo fmt and
cargo clippy --all-targets --all-features -- -D warnings are clean.

Closes #711
